WIP: refactoring `svm-ffi` tracking module.
